requirements.txt
flask
flask-cors
requests
openai
yt-dlp
git+https://github.com/openai/whisper.git
torch
ffmpeg-python
transformers
sentencepiece
python-dotenv